ucentral: development update
* update netifd
* update hostapd
* update mt76
* update maverick
* add dynamic vlan
* update ucentral-client
* update ucentral-schema
* update ucode
* update 5.10 patches
* fix realtek

+++ feeds/tip/maverick/files/usr/libexec/ucentral/maverick.sh
@@ -1,5 +1,9 @@
 #!/bin/sh
 
+active=$(ubus call ucentral status | jsonfilter -e '@.active')
+
+[ -n "$active" -a ! "$active" -eq 1 ] && exit 0
+
 rm /etc/config/network /etc/config/wireless
 cp /rom/etc/config/uhttpd /rom/etc/config/firewall /rom/etc/config/dhcp /rom/etc/config/dropbear /etc/config
 config_generate
